+++ src/pages/index.tsx
@@ -1,3 +1,5 @@
-const HomePage = () => {};
+const HomePage = () => {
+  return <></>;
+};
 
 export default HomePage;
