+++ src/styles/globals.css
@@ -28,7 +28,8 @@
   --gray-500: #474d66;
   --gray-600: #3b415b;
 
-  --green-100: #eef9f6;
+  --green-50: #eef9f6;
+  --green-100: #6cd9bf;
   --green-200: #4cbfa4;
   --green-300: #32a68a;
